HAWQ-1644. Make delegation token optional in PXF
Make delegation token property optional for PXF when used against secure
hadoop. Delegation token is a complementary means for authentication
with hadoop along with Kerberos. There is not much value in using the
delegation token with masterless and segment only PXF architecture, as
each pxf jvm will need to establish the token anyway. Simply
authenticating with Keberos should be good enough in such a deployment
mode.

Co-authored-by: Alexander Denissov <adenissov@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>
Co-authored-by: Shivram Mani <smani@pivotal.io>